+++ src/steps/intune/constants.ts
@@ -142,8 +142,16 @@ export const relationships = {
       _class: RelationshipClass.ASSIGNED,
     },
   ),
-  MULTI_DEVICE_HAS_DETECTED_APPLICATION: createRelationshipForAllDeviceTypes({
+  MULTI_DEVICE_INSTALLED_DETECTED_APPLICATION: createRelationshipForAllDeviceTypes(
+    {
+      targetType: entities.DETECTED_APPLICATION._type,
+      _class: RelationshipClass.INSTALLED,
+    },
+  ),
+  MANAGED_APPLICATION_MANAGES_DETECTED_APPLICATION: {
+    _type: 'intune_managed_application_manages_detected_application',
+    sourceType: entities.MANAGED_APPLICATION._type,
+    _class: RelationshipClass.MANAGES,
     targetType: entities.DETECTED_APPLICATION._type,
-    _class: RelationshipClass.HAS,
-  }),
+  },
 } as const;